Rename user to lowercase; makes more sense
